Add files via upload
Study + template examples. The main objective of this research is to investigate the user-centered design (UCD), specifically user personas, as methodological tools to inform the design and development of Consumer Health Tools (CHT) technologies for an aging population. This study took place in China. It concludes user personas have value as they capture the conceptual model of the aging and informing the design and development decisions of CHT.

Authors share their methodology and examples of of personas created (Urban v Rural)